+++ ketos/__init__.py
@@ -1,5 +1,1 @@
-#import ketos.data_handling
-#import ketos.audio
-#import ketos.neural_networks
-
-#__all__ = ["data_handling", "audio", "neural_networks"]
+__version__ = "2.5.1"